date separator done
