--- src/components/DateSeparator/index.js
@@ -1,1 +1,0 @@
-export { default as DateSeparator } from './DateSeparator';
